Stm/StmPkg/Core/Runtime - Add execution processor number to shared header
This patch adds the VM/PE execution processor number to the STM
shared header.

It also formalizes the fields of the shared STM header as this is the
third element to be added. Before, the header definition was informal
where each element was identified as an offset.

This patch also cleans up some of the code related to the header in
PeSmiHandler.c by removing duplicative code.

Signed-off-by: Eugene Myers <edmyers@tycho.nsa.gov>