Core/Init/StmInit.c - Clear possible VMCS cache residue
Make sure that cache VMCS residue is cleared from the processor.

This can affect the STM VM's when the STM is torn down when the
proessor is suspended.  In this instance the processor may hold
in it's VMCS cache the previous SMI and SMM VMCS's, which if not
cleared can corrupt the new VMCS and cause unpredicatable actions
to occur.

Signed-off-by: Eugene Myers <edmyers@tycho.nsa.gov>