Start VM/PE frpm non-VM/PE processor
Current implementation of STM/PE requires that the VM/PE and any
run commands that affect the VM/PE be done on the VM/PE processor.
On most operating systems, the VM/PE processor has been hot
unplugged from the available processor set.  Getting commands to
the VM/PE requires some modification of the operating system for
this action to be supported and this implemantation is fragile
across releases.

This patch allows VM/PE commands to issued via a vmcall on any
processor.  The STM will setup a SMI timer interrupt to cause
the command to be passed to the VM/PE processor.  This greatly
reduces and simplifies the changes needed for the O/S.

Signed-off-by: Eugene Myers <edmyers@cyberpackventures.com>